requirements.txt
beautifulsoup4==4.12.0
blinker==1.6.2
blis==0.7.11
catalogue==2.0.10
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.3
cymem==2.0.8
deep-translator==1.11.4
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.0/en_core_web_lg-3.7.0-py3-none-any.whl
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl
filelock==3.10.7
Flask==2.3.3
gensim==4.3.2
google-play-scraper==1.2.4
idna==3.4
instagrapi==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
langcodes==3.3.0
lxml==4.9.2
MarkupSafe==2.1.3
murmurhash==1.0.10
mysql-connector==2.2.9
nltk==3.8.1
numpy==1.24.2
packaging==23.2
pandas==2.0.0
Pillow==10.0.0
preshed==3.0.9
pycryptodomex==3.18.0
pydantic==1.10.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
regex==2023.10.3
requests==2.28.2
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
sklearn==0.0.post10
smart-open==6.4.0
snscrape==0.6.2.20230320
soupsieve==2.4
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
thinc==8.2.1
threadpoolctl==3.2.0
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.15
wasabi==1.1.2
weasel==0.3.3
Werkzeug==2.3.7
